started retroactive compensation refactor
