+++ packages/i18n/locales/en/translation.json
@@ -384,7 +384,6 @@
     "alreadySentTokenSwap_gov": "The chain has already sent its share of this token swap.",
     "alreadySentTokenSwap_wallet": "You have already sent your share of this token swap.",
     "browserNotSupported": "Not supported in this browser.",
-    "cannotCreateCompensationCycleAlreadyActive": "You cannot create a new compensation cycle while one is already active.",
     "cannotStakeMoreThanYouHave": "You can't stake or unstake more tokens than you have.",
     "cannotTxZeroTokens": "You can't stake, unstake, or claim zero tokens.",
     "cannotUseCreateNftCollectionOnStargaze": "You cannot create a new NFT Collection via this action on Stargaze. Use the Stargaze Studio app on the DAO's Apps page instead.",
