+++ packages/stateful/widgets/widgets/RetroactiveCompensation/Renderer/components/stateless/TabRenderer.tsx
@@ -1,94 +1,60 @@
-import { Add, ArrowDropDown, Ballot, Remove } from '@mui/icons-material'
-import clsx from 'clsx'
-import { ComponentType, useState } from 'react'
+import { Add, ChevronLeft, Remove } from '@mui/icons-material'
+import { ComponentType } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import {
-  Button,
-  LineLoader,
-  LineLoaders,
-  NoContent,
   Tooltip,
   useDaoInfoContext,
   useDaoNavHelpers,
 } from '@dao-dao/stateless'
-import { LoadingData, LoadingDataWithError } from '@dao-dao/types'
-import { IconButtonLinkProps } from '@dao-dao/types/components/IconButtonLink'
+import { ButtonLinkProps, WidgetId } from '@dao-dao/types'
 
-import {
-  CompletedSurveyListing,
-  StatefulOpenSurveySectionProps,
-  Status,
-} from '../../types'
-import { CompletedSurveyRow } from './CompletedSurveyRow'
+import { PagePath } from '../../types'
 
 export interface TabRendererProps {
-  loadingStatus: LoadingData<Status | undefined>
-  loadingCompletedSurveys: LoadingData<CompletedSurveyListing[]>
   isMember: boolean
-  NewSurveyForm: ComponentType
-  OpenSurveySection: ComponentType<
-    Pick<StatefulOpenSurveySectionProps, 'status'>
-  >
-  downloadCompletedSurvey: (pastSurvey: CompletedSurveyListing) => void
-  loadingCompletedSurveyId: number | undefined
-  loadingMembershipDuringCompletedSurveys: LoadingDataWithError<string[]>
-  IconButtonLink: ComponentType<IconButtonLinkProps>
+  pages: Record<PagePath, ComponentType>
+  ButtonLink: ComponentType<ButtonLinkProps>
 }
 
 export const TabRenderer = ({
-  loadingStatus,
-  loadingCompletedSurveys,
   isMember,
-  NewSurveyForm,
-  OpenSurveySection,
-  downloadCompletedSurvey,
-  loadingCompletedSurveyId,
-  loadingMembershipDuringCompletedSurveys,
-  IconButtonLink,
+  pages,
+  ButtonLink,
 }: TabRendererProps) => {
   const { t } = useTranslation()
   const { coreAddress } = useDaoInfoContext()
-  const { getDaoProposalPath } = useDaoNavHelpers()
+  const { daoSubpathComponents, getDaoPath } = useDaoNavHelpers()
 
-  const [showCreate, setShowCreate] = useState(false)
-  // Can create survey if member of DAO and there does not exist a current
-  // survey.
-  const canCreateSurvey =
-    isMember && !loadingStatus.loading && !loadingStatus.data
-
-  // Subtitle status.
-  const subtitleStatus =
-    !loadingStatus.loading && !loadingStatus.data
-      ? t('info.noActiveCompensationCycle')
-      : null
+  const pagePath = daoSubpathComponents[1] || ''
+  const Page =
+    pagePath in pages ? pages[pagePath as PagePath] : pages[PagePath.Home]
 
   return (
     <div className="flex flex-col gap-6">
       <div className="flex flex-row items-center justify-between gap-8 border-b border-border-secondary pb-6">
-        <div className="flex flex-row flex-wrap items-center gap-x-4 gap-y-1">
-          <p className="title-text text-text-body">
-            {t('title.retroactiveCompensation')}
-          </p>
-
-          {subtitleStatus && <p className="secondary-text">{subtitleStatus}</p>}
-        </div>
+        <p className="title-text text-text-body">
+          {t('title.retroactiveCompensation')}
+        </p>
 
         <Tooltip
           title={
             !isMember
               ? t('error.mustBeMemberToCreateCompensationCycle')
-              : !loadingStatus.loading && loadingStatus.data
-              ? t('error.cannotCreateCompensationCycleAlreadyActive')
               : undefined
           }
         >
-          <Button
+          <ButtonLink
             className="shrink-0"
-            disabled={!canCreateSurvey}
-            onClick={() => setShowCreate((c) => !c)}
+            disabled={!isMember || pagePath === PagePath.Create}
+            href={getDaoPath(
+              coreAddress,
+              WidgetId.RetroactiveCompensation +
+                '/' +
+                (pagePath === PagePath.Create ? PagePath.Home : PagePath.Create)
+            )}
           >
-            {canCreateSurvey && showCreate ? (
+            {pagePath === PagePath.Create ? (
               <>
                 <Remove className="!h-4 !w-4" />
                 {t('button.cancel')}
@@ -102,86 +68,21 @@ export const TabRenderer = ({
                 <span className="md:hidden">{t('button.new')}</span>
               </>
             )}
-          </Button>
+          </ButtonLink>
         </Tooltip>
       </div>
 
-      {loadingStatus.loading ? (
-        <LineLoader type="retroactive" />
-      ) : // If no active survey, text is shown at the top. No need to render anything here.
-      !loadingStatus.data ? null : (
-        <OpenSurveySection status={loadingStatus.data} />
+      {/* Back button. */}
+      {pagePath !== PagePath.Home && (
+        <ButtonLink
+          href={getDaoPath(coreAddress, WidgetId.RetroactiveCompensation)}
+        >
+          <ChevronLeft className="!h-4 !w-4" />
+          {t('button.back')}
+        </ButtonLink>
       )}
 
-      {canCreateSurvey && showCreate ? (
-        <NewSurveyForm />
-      ) : (
-        <>
-          <div className="link-text ml-2 flex flex-row items-center gap-3">
-            <ArrowDropDown className="!h-4 !w-4 text-icon-primary" />
-            <p className="text-text-secondary">{t('title.history')}</p>
-          </div>
-
-          {loadingCompletedSurveys.loading ? (
-            <LineLoaders lines={20} type="retroactive" />
-          ) : loadingCompletedSurveys.data.length > 0 ? (
-            <div className="flex flex-col gap-1">
-              {loadingCompletedSurveys.data.map((survey, index) => {
-                const wasMemberDuringSurvey =
-                  !loadingMembershipDuringCompletedSurveys.loading &&
-                  !loadingMembershipDuringCompletedSurveys.errored &&
-                  Number(loadingMembershipDuringCompletedSurveys.data[index]) >
-                    0
-
-                return (
-                  <CompletedSurveyRow
-                    key={survey.id}
-                    IconButtonLink={IconButtonLink}
-                    className={clsx(
-                      // If survey is loading, animate pulse.
-                      loadingCompletedSurveyId === survey.id && 'animate-pulse',
-                      // If was not a member and no proposal, disable pointer
-                      // events since we can't do anything.
-                      !wasMemberDuringSurvey &&
-                        !survey.proposalId &&
-                        'pointer-events-none'
-                    )}
-                    onClick={() =>
-                      wasMemberDuringSurvey
-                        ? // If was a member, prompt for authentication before downloading CSV.
-                          downloadCompletedSurvey(survey)
-                        : // If was not a member but proposal exists, open survey in new tab on select.
-                          survey.proposalId &&
-                          window.open(
-                            getDaoProposalPath(coreAddress, survey.proposalId),
-                            '_blank'
-                          )
-                    }
-                    survey={survey}
-                    tooltip={
-                      wasMemberDuringSurvey
-                        ? t('button.downloadSurveyCsv')
-                        : t('button.goToProposal')
-                    }
-                  />
-                )
-              })}
-            </div>
-          ) : (
-            <NoContent
-              Icon={Ballot}
-              actionNudge={t('info.createFirstOneQuestion')}
-              body={t('info.noCompletedCyclesYet')}
-              buttonLabel={t('button.newCompensationCycle')}
-              onClick={
-                canCreateSurvey && !showCreate
-                  ? () => setShowCreate(true)
-                  : undefined
-              }
-            />
-          )}
-        </>
-      )}
+      <Page />
     </div>
   )
 }
